api route testing done for now

+++ services/authentication/src/api/user/controller.ts
@@ -18,7 +18,6 @@ import { configs } from '~config';
 
 const {
   COOKIE_AUTH_KEY_NAME: authKeyName,
-  COOKIE_CSRF_KEY_NAME: csrfKeyname,
 } = configs;
 
 export const requestUserInfo = async (
@@ -67,69 +66,7 @@ export const requestUserInfo = async (
     );
 
   } catch (err) {
-    return error(
-      res,
-      {
-        ...err,
-        message: 'Internal Server Error',
-      },
-    );
-  }
-};
-
-export const verify = async (
-  req: Request,
-  res: Response,
-): Promise<Response> => {
-  try {
-    const {
-      exclude_fields = [],
-      search_values = [],
-      search_field = 'id',
-      return_fields = [],
-    } = req.query as QueryParamsObject;
-
-    if (!UserModel.ALL_FIELDS.includes(search_field)) {
-      return badRequest(
-        res,
-        `${search_field} is not a valid field.`,
-      );
-    }
-
-    const returnFields = Array.from(
-      new Set([
-        ...UserModel.BASE_FIELDS,
-        ...return_fields.filter((return_field) => {
-          return UserModel.ALL_FIELDS.includes(return_field)
-            && !exclude_fields.includes(return_field);
-        }),
-      ]),
-    );
-
-    const users: UserModel[] = await UserModel
-      .query()
-      .where(
-        search_field,
-        'in',
-        search_values,
-      )
-      .pick(returnFields);
-
-    return success(
-      res,
-      {
-        users,
-      },
-    );
-
-  } catch (err) {
-    return error(
-      res,
-      {
-        ...err,
-        message: 'Internal Server Error',
-      },
-    );
+    return error(res, err);
   }
 };
 
@@ -137,29 +74,30 @@ export const logout = async (
   req: UserInfoRequest<UserModel>,
   res: Response,
 ): Promise<Response> => {
-  res.clearCookie(authKeyName);
-  res.clearCookie(csrfKeyname);
-
-  const user_id = req.user.id;
-  const refresh_token = req.refresh_token;
-
   try {
+    const {
+      user: {
+        id: userId,
+      },
+      refresh_token,
+    } = req;
+
+    if (!userId || !refresh_token) {
+      return badRequest(res);
+    }
 
     await TokenModel
       .query()
       .delete()
-      .where('user_id',  user_id)
+      .where('user_id',  userId)
       .where('refresh_token', refresh_token);
 
-    return res.send(204);
+    return res
+      .clearCookie(authKeyName)
+      .status(204)
+      .send();
 
   } catch (err) {
-    return error(
-      res,
-      {
-        ...err,
-        message: 'Internal Server Error',
-      },
-    );
+    return error(res, err);
   }
 };
